pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-kb-quiz"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.40.0",
    "httpx>=0.27.0",
    "typer>=0.12.0",
    "PyYAML>=6.0",
    "sentence-transformers>=3.0.0",
    "chromadb>=0.6.0",
    "RestrictedPython>=7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-mock>=3.12.0",
    "respx>=0.21.0",
]

[project.scripts]
ai-kb-quiz = "cli.main:app"

[tool.setuptools.packages.find]
include = ["engine*", "cli*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "local: requires real local dependencies (Ollama, ChromaDB); skipped in CI",
]
